Fix `out_of_range` error in mappings
The `out_of_range` error expects a string argument. To get it we were converting the input value to string using our `to_string` method which in turn will throw an error because the value isn't valid. 

Instead we cast it to the udnerlying type and convert that to string.